Add collection image attribution information
Adding in image attribution into the CollectionInfo API,
and updated all the existing collections with this info.
Replacing coin images just to ensure the attribution is
correct.  I still need to add code that calls this
method and aggregates the information for displaying on
the info page.

Fixing a bug sw23 identified in the logic to detect
whether no mint marks have been selected when creating
a new collection.