Refactor by seperating out the core app code
Modifying the directory structure to eventually mirror how
the repo will be structured - the core code contained in one
standalone package, and the app-specific code outside of it.
The next steps are to move the resource files into here as
well (unfortunately I can't think of a good way to keep them
while moving to this model) along with the strings.